Admin portal

+++ MenrvaBooks/src/main/kotlin/com/menrva/repositories/SeriesRepository.kt
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface SeriesRepository : JpaRepository<Series, Int> {
+interface SeriesRepository : JpaRepository<Series, Long> {
     fun findByName(name: String): List<Series>
 }
